Update README.md
# SCY-FHY
The log-in page of a science fiction webpage enhanced with dif
The page has two different layouts,one that is triggered when the width of the screen is greater than the height and the other, when screen-height is greater than sreen-width.
Every object on the screen would be sized relative to the size of the screen. The page had been built in a flexible way to allow its contents to be manipulated by people ranging from JavaScript geeks to people who have little or no knowledge about JavaScript.